remaps now rendered as augeas hashes, not arrays
rather than always rendering remaps as array, and hence after a chain of
changes ending up with *one* line per map, we now render it as hash,
which makes it easy to add new or remove old rules, by simply refering
to it via the URL which is being remaped.

We also switch from using an inlinetemplate() hack to simply exposing a
Ruby Hash method.