use property instead param
parameters are not something that self.instances can discover or fill,
we definitely always want properties for everything that we can discover
from the system.

parameters are flags that influence the behaviour, or things that cannot
be reverse-engineered, like passwords (we set a password, it's converted
to hash in the system, but we can't get discover the password, hence
new param in that case)

name is different, in this regard, but probably only because it's a
namevar in this case.