disregard previous comment about idempotency...
simplify (copy-paste) self.prefetch from @glarizza's
http://garylarizza.com/blog/2013/12/15/seriously-what-is-this-provider-doing/
because the one in puppet types & providers doesn't actually work:

it appears that it breaks down when comparing `is` & `should`.

*now* it actually works. the previous commit had a test in value=()
which was nonsense: value= has no access to the previous value.

if none of the above makes sense, that's because it only kinda makes
half-sense to me... btw, new best friend:

    require 'pry' ; binding.pry

big thanks to @apenny for this one.