third time is a charm, again ;)
switch implementation of trafficserver_storage to @adrienthebo's
filemapper. we keep our regex simple: strip leading and trailing spaces.
then just use split to.. split the values.

please note that right now we are still running into a mysterkious
issue with the validate block in the type. it appears that not all
values are treated equally: properties don't get a value, even though
it exists in @property_hash, and can indirectly be accessed or seen in
`self`.