do not (re)set other values in setter
in #value= we used to re-set @property_hash[:name], this shouldn't be
necessary. the only thing needed is @property_hash[:value], after all,
we're the `value` setter.

further, make sure to return :record in `initialize`, and set it to the
same value as :name. after all `:record` is the cannonical `name`.