notify the service instead of running traffic_line
traffic_line has been deprecated, so it would be tricky to emulate the
same behaviour with traffic_ctl, especially since there are versions of
trafficserver that have *both* binaries.

Instead of trying to be clever, we'll simply notify the service class.

(also: fix a small syntax issue in the README!)